+++ softlight/tests/sl_sdf_image_test.cpp
@@ -182,7 +182,7 @@ utils::Pointer<SL_SceneGraph> mesh_test_create_context()
     retCode = meshLoader.load("testdata/towerG.obj", opts);
     assert(retCode != 0);
 
-    retCode = pGraph->import(meshLoader.data());
+    retCode = (int)pGraph->import(meshLoader.data());
     assert(retCode == 0);
 
     // Always make sure the scene graph is updated before rendering
